@@ -1,16 +1,16 @@
 name: klutter
 description: Write Flutter plugins for both Android and iOS using Kotlin only.
-version: 1.0.0
+version: 2.0.0
 homepage: https://buijs.dev
 repository: https://github.com/buijs-dev/klutter-dart
 
 environment:
-  sdk: ">=2.15.1 <4.0.0"
+  sdk: ">=2.17.6 <4.0.0"
 
 dependencies:
-  dart_style: ^2.2.3
+  dart_style: ^2.2.4 # can be bumped when min SDK is 2.19.0
 
 dev_dependencies:
-  coverage: ^1.3.2
-  dartdoc: ^6.2.1
-  test: ^1.19.5
+  coverage: ^1.6.4
+  dartdoc: ^6.1.0 # can be bumped when min SDK is 2.19.0
+  test: ^1.24.1
