@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2022 Buijs Software
+// Copyright (c) 2021 - 2023 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import "cli.dart";
+import "../../klutter.dart";
 
 /// Object to store parsed user input.
 class Command {
@@ -29,14 +29,14 @@ class Command {
   const Command({
     required this.taskName,
     required this.scriptName,
-    this.option,
+    this.options = const <ScriptOption, String>{},
   });
 
   /// Name of Task to be executed.
   final TaskName taskName;
 
-  /// Optional value which may or may not be used by the Task.
-  final String? option;
+  /// Optional values which may or may not be used by the Task.
+  final Map<ScriptOption, String> options;
 
   /// Name of script which should contain a Task with [TaskName].
   final ScriptName scriptName;
@@ -48,29 +48,115 @@ class Command {
   /// or null if the command (user input) is invalid.
   static Command? from({
     required ScriptName script,
-    required String task,
+    required List<String> arguments,
   }) {
-    /// Regex to parse the CLI arguments.
-    final taskRegex = RegExp(
-      r"^\s*(init|add|install)\s*=*\s*(\S+|$)",
-    );
-
-    final match = taskRegex.firstMatch(task);
-
-    if (match == null) {
+    if (arguments.isEmpty) {
       return null;
     }
 
-    final taskName = match.group(1).toTaskNameOrNull;
+    final taskName = arguments.first.toTaskNameOrNull;
 
     if (taskName == null) {
       return null;
     }
 
-    return Command(
-      option: match.group(2),
-      taskName: taskName,
-      scriptName: script,
-    );
+    final args = arguments.sublist(1, arguments.length);
+
+    if (script == ScriptName.consumer) {
+      if (taskName == TaskName.init) {
+        if (args.isEmpty) {
+          return Command(taskName: taskName, scriptName: script);
+        } else {
+          return null;
+        }
+      }
+
+      if (taskName == TaskName.add) {
+        if (args.length != 1) {
+          return null;
+        }
+
+        final option = args.first.split("=");
+
+        if (option.length != 2) {
+          return null;
+        }
+
+        if (option.first != ScriptOption.lib.name) {
+          return null;
+        }
+
+        return Command(
+            taskName: taskName,
+            scriptName: script,
+            options: {ScriptOption.lib: option[1]});
+      }
+    }
+
+    if (script == ScriptName.producer) {
+      if (taskName == TaskName.init) {
+        if (args.length > 2) {
+          return null;
+        }
+
+        final options = <ScriptOption, String>{};
+
+        for (final arg in args) {
+          final option = arg.split("=");
+
+          if (option.length != 2) {
+            return null;
+          }
+
+          if (option.first == ScriptOption.bom.name) {
+            options[ScriptOption.bom] = option[1];
+          } else if (option.first == ScriptOption.flutter.name) {
+            options[ScriptOption.flutter] = option[1];
+          } else {
+            return null;
+          }
+        }
+
+        if (!options.containsKey(ScriptOption.bom)) {
+          options[ScriptOption.bom] = klutterGradleVersion;
+        }
+
+        if (!options.containsKey(ScriptOption.flutter)) {
+          options[ScriptOption.flutter] = klutterFlutterVersion;
+        }
+
+        return Command(
+            taskName: taskName, scriptName: script, options: options);
+      }
+
+      if (taskName == TaskName.get) {
+        if (args.length != 1) {
+          return null;
+        }
+
+        final option = args.first.split("=");
+
+        if (option.length != 2) {
+          return null;
+        }
+
+        if (option.first != ScriptOption.flutter.name) {
+          return null;
+        }
+
+        final version = option[1].verifyFlutterVersion?.version;
+
+        if (version == null) {
+          return null;
+        }
+
+        return Command(
+            taskName: taskName,
+            scriptName: script,
+            options: <ScriptOption, String>{ScriptOption.flutter: version});
+      }
+    }
+
+    return null;
   }
 }
