@@ -72,13 +72,11 @@ void createPlatformModule({
   required String packageName,
 }) =>
     PlatformModule.fromRoot(
-      pathToRoot: pathToRoot,
-      pluginName: pluginName,
-      packageName: packageName,
-    )
+        pathToRoot: pathToRoot,
+        pluginName: pluginName,
+        packageName: packageName)
       ..createPlatformGradleFile
       ..createPlatformSourceFolders
-      ..createAndroidManifest
       ..createAndroidPlatformClass
       ..createCommonGreetingClass
       ..createCommonPlatformClass
@@ -114,7 +112,7 @@ extension on File {
   /// Write the content of the settings.gradle.kts of a Klutter plugin.
   void writeSettingsGradleContent(String pluginName) {
     writeAsStringSync('''
-            // Copyright (c) 2021 - 2022 Buijs Software
+            // Copyright (c) 2021 - 2023 Buijs Software
             |//
             |// Permission is hereby granted, free of charge, to any person obtaining a copy
             |// of this software and associated documentation files (the "Software"), to deal
@@ -152,8 +150,8 @@ extension on File {
           |        maven { url = uri("https://repsy.io/mvn/buijs-dev/klutter") }
           |    }
           |    dependencies {
-          |        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10")
-          |        classpath("com.android.tools.build:gradle:7.0.4")
+          |        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:1.8.20")
+          |        classpath("com.android.tools.build:gradle:8.0.2")
           |        classpath(platform("dev.buijs.klutter:bom:$klutterBomVersion"))
           |        classpath("dev.buijs.klutter:gradle")
           |    }
@@ -273,16 +271,10 @@ class PlatformModule {
       |
       |    include("bill-of-materials")
       |}
-      |    
-      |ksp {
-      |    arg("klutterScanFolder", project.buildDir.absolutePath)
-      |    arg("klutterOutputFolder", project.projectDir.parentFile.absolutePath)
-      |    arg("klutterGenerateAdapters", "true")
-      |    arg("intelMac", "false") // Set to "true" if you're building on an Intel Mac!
-      |}
       |
       |kotlin {
       |
+      |    jvmToolchain(17)
       |    android()
       |
       |    val xcfName = "Platform"
@@ -309,7 +301,7 @@ class PlatformModule {
       |        val commonMain by getting {
       |            dependencies {
       |                implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.5.0")
-      |                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.0")
+      |                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4")
       |            }
       |        }
       |
@@ -358,12 +350,15 @@ class PlatformModule {
       |}
       |
       |android {
-      |    compileSdk = $androidCompileSdk
-      |    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
+      |    namespace = "$packageName.platform"
       |    sourceSets["main"].kotlin { srcDirs("src/androidMain/kotlin") }
+      |    compileOptions {
+      |        sourceCompatibility = JavaVersion.VERSION_17
+      |        targetCompatibility = JavaVersion.VERSION_17
+      |    }
       |    defaultConfig {
+      |        compileSdk = $androidCompileSdk
       |        minSdk = $androidMinSdk
-      |        targetSdk = $androidTargetSdk
       |    }
       |}
       |
@@ -376,23 +371,6 @@ class PlatformModule {
       |    .setFinalizedBy(listOf(tasks.getByName("klutterCopyFramework")))
       |"""
           .format);
-  }
-
-  /// Create the AndroidManifest.xml file.
-  ///
-  /// Will create a new file if it does not exist
-  /// or overwrite the current AndroidManifest.xml
-  /// if it already exists.
-  void get createAndroidManifest {
-    root.resolveFile("src/androidMain/AndroidManifest.xml").normalizeToFile
-      ..maybeCreate
-      ..writeAsStringSync(
-        """
-        <?xml version="1.0" encoding="utf-8"?>
-        |<manifest package="$packageName.platform" />
-        """
-            .format,
-      );
   }
 
   /// Create Platform.kt (Kotlin) class file
