@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2022 Buijs Software
+// Copyright (c) 2021 - 2023 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -28,8 +28,7 @@ void main() {
       " add ": TaskName.add,
       " init": TaskName.init,
       "INIT": TaskName.init,
-      "install": TaskName.install,
-      "i n s t a l l": null,
+      "i n i t": null,
       "destroyAllHumans!": null,
       null: null,
       "": null,
